Comment controllers are corrected some extent

+++ src/controllers/comment.controller.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import { Comment } from "../models/comment.model.js";
 import { ApiError } from "../utils/apiError.js";
 import { ApiResponse } from "../utils/apiResponse.js";
@@ -6,27 +7,29 @@ import { asyncHandler } from "../utils/asyncHandler.js";
 const getVideoComments = asyncHandler(async (req,res)=>{
     // ! May change later
 
-    const {video} = req.params.video // here video is a complete obj not just id
+    const video = req.params.video // here video is a complete obj not just id
     if(!video){
         throw ApiError(400,"Video does not exist")
     }
     
-    const comments = await Comment.find({video:video._id})
-    if(!comments){
+    const commentObjArray = await Comment.find({video:video._id})
+    if(!commentObjArray){
         throw ApiError(400,"No comments present in the video")
     }
-    
+        
     // Aggregate funciton to be used for fetching the comments
-    
+    res.status(200).json(
+        new ApiResponse(200,commentObjArray,"All Comments fetched successfully!")
+    )
 })
 
 const createComment = asyncHandler(async (req,res)=>{
 
     //! May change later
 
-    const {user} = req.user
-    const {video} = req.video
-    const {commentContent} = req.body.commentContent
+    const user = req.user
+    const video = req.video
+    const commentContent = req.body.commentContent
 
     if(!user){
         throw ApiError(401,"Unauthorized access to comment on video")
@@ -43,7 +46,7 @@ const createComment = asyncHandler(async (req,res)=>{
         content:commentContent,
         video: video._id // ! May change later
     })
-    
+
     res.status(200).json(
         new ApiResponse(200,comment,"Commented on the video successfully...")
     )
@@ -52,10 +55,11 @@ const createComment = asyncHandler(async (req,res)=>{
 const deleteComment = asyncHandler(async (req,res)=>{
 
     //! May change later
-
-    const {user} = req.user
-    const {video} = req.video
-    const {comment} = req.comment
+    //! assuming the front end will provide the id
+    //! assuming delete option will not be shown to other users rather than commentee himself
+    const user = req.user
+    const video = req.video
+    const comment = req.comment // complete comment obejct
 
     if(!user){
         throw ApiError(401,"Unauthorized access to comment on video")
@@ -67,27 +71,33 @@ const deleteComment = asyncHandler(async (req,res)=>{
         throw ApiError(401,"comment is not getting passed")
     }
 
-    let deleteComment;
-    try {
-        deleteComment = await Comment.findByIdAndDelete(comment._id)
-    } catch (err) {
-        throw new ApiError(400,"Comment does not exist")
+    if(comment.owner !== user._id){
+        throw new ApiError(400,"Unauthorized accesss to delete commetn")
     }
 
-    console.log(deleteComment)
-
-    res.status(200).json(
-        new ApiResponse(200,deleteComment,"Comment has been deleted successfully...")
-    )
+    try {
+        const deleteComment = await Comment.findByIdAndDelete(comment._id)
+        
+        console.log(deleteComment)
+        if(!deleteComment){
+            throw new ApiError(401,"Somehting went wrong ");
+        }
+        res.status(200).json(
+            new ApiResponse(200,deleteComment,"Comment has been deleted successfully...")
+        )
+    } catch (err){
+        throw new ApiError(400,"Comment could not be deleted")
+    }
 })
+
 const updateComment = asyncHandler(async (req,res)=>{
 
     //! May change later
 
-    const {user} = req.user
-    const {video} = req.video
-    const {comment} = req.comment
-    const {newcomment} = req.body.newcomment
+    const user = req.user
+    const video = req.video
+    const comment = req.comment
+    const newcomment = req.body.newcomment
 
     if(!user){
         throw ApiError(401,"Unauthorized access to comment on video")
@@ -116,8 +126,9 @@ const updateComment = asyncHandler(async (req,res)=>{
     existingComment.content = newcomment
     existingComment.save({validateBeforeSave: false});
 
+    const updatedCommentRes = Comment.findById(comment._id)
     res.status(200).json(
-        new ApiResponse(200,updateComment,"Comment has been updated successfully...")
+        new ApiResponse(200,updatedCommentRes,"Comment has been updated successfully...")
     )
 })
